# doc/api.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://localhost:3000
info:
  version: 1.0.0
  title: WasaPhoto application project
  description: >-
    This document describes the API for WasaPhoto project
    
    'Keep in touch with your friends by sharing photos of special moments,
    thanks to WASAPhoto! Directly from your PC,
    you can upload your photos, and they will be visible to everyone
    who is following you.'
    
  contact:
    name: "Olha Biziura"
    email: "olhabiziura@gmail.com"
...
paths:
...
  /deletephoto/{pictureID}:
    delete:
      operationId: deletePhoto
      tags:
        - photos
      summary: Deletes a photo
      description: Deleting a photo by its ID.
      parameters:
        - in: path
          name: picture_id
          required: true
          schema:
            type: integer  # Adjust schema if picture_id is not an integer
          description: The numeric ID of the photo to be deleted.
      responses:
        '204':
          description: The photo has been deleted successfully
        '400':
          description: The request was not valid. Can be invalid user_id or picture_id.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...